Added SERVER.BIN file to SDcard_* direcotries. Only a server will read these, to get its SmartHome Node ID info The Linux server app using wxWidggets is benning to have a functional GUI. The "wall control" buttons are responding and giving debug info, and also are appropriately controlling the current level guage. Linux app development so far is done in Lubuntu virtual machine within a VMware player, currently on my work laptop.
room and load buttons do NOT yet update with names, or update load info

text log does NOT yet work/not implemented
Zigbee does NOT yet work/not implemented
ethernet/internet does NOT yet work/not implemented

+++ arduino/SmartHomeProject/wallControl/wallControl.ino
@@ -322,7 +322,7 @@ uint16_t getLoadNodeIDfromSD(uint16_t roomNum, uint8_t loadNum)
 
 
 // Get SmartHome Load Type value for this Wall Control unit from SD card
-uint16_t getLoadTypefromSD(uint16_t roomNum, uint8_t loadNum)
+uint8_t getLoadTypefromSD(uint16_t roomNum, uint8_t loadNum)
 {
     uint8_t tmpLoadType = NODEINFO_NODETYPE_LIGHT; //loads are not controls, and more loads will be lights than fans
     char loadNumFileName[16] = "/65535/255.BIN";  // default longest filename to make sure have enough chars in string
